+++ Tablero/Clasic YuGiOh!/src/mainPrincipal/Main.java
@@ -17,7 +17,6 @@ public class Main {
          
 		System.out.println(tableroController.getManoMonstruoOponente());
 
-	 
 	
 	}
